Make page object helpers resuable
