+++ tests/helpers/make-component.js
@@ -1,0 +1,10 @@
+// Make a basic page object component and merge in options
+const makeComponent = (selector, options) => {
+  let baseComponent = {
+    scope: selector
+  };
+
+  return Object.assign({}, baseComponent, options);
+}
+
+export default makeComponent;
